photo-mapper/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0
pandas==2.1.3